% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{coronavirus_br_states}
\alias{coronavirus_br_states}
\title{The Coronavirus COVID-19 Brazil Dataset by state}
\format{A data.frame object
\describe{
  \item{date}{Date}
  \item{state}{Brazilian state}
  \item{cases}{Confirmed Cases}
  \item{deaths}{Deaths}
  \item{delta_cases}{Confirmed Cases daily variation}
  \item{delta_deaths}{Deaths daily variation}
  \item{days_gt_10}{Days since 10th Confirmed Case}
  \item{delta_deaths}{Days since 100th Confirmed Case}
}}
\source{
\href{https://covid.saude.gov.br}{Ministerio da Saude} and \href{https://brasil.io/dataset/covid19/}{brasil.io}
}
\usage{
coronavirus_br_states
}
\description{
Daily summary of the Coronavirus (COVID-19) cases in Brazil by state
}
\details{
The dataset contains the daily summary of Coronavirus cases (confirmed cases and deaths) in brazil
}
\examples{

data(coronavirus_br_states)

}
\keyword{COVID19}
\keyword{coronavirus_br}
\keyword{datasets}
