% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{coronavirus_rj_case_metadata}
\alias{coronavirus_rj_case_metadata}
\title{Rio de Janeiro COVID-19 cases}
\format{data.frame object
\describe{
  \item{bairro_resid__estadia}{Rio de Janeiro Neighborhood}
  \item{classificação_final}{Final Classification}
  \item{dt_notific}{Notification date}
  \item{ap__residencia}{AP of the patient residence}
  \item{idade}{age}
  \item{sexo}{sex}
  \item{ObjectId2}{objectId2}
  \item{sexo}{sex}
  \item{faixa_etária}{age range}
  \item{hospitalizações}{hospitalization}
  \item{uti}{Intensive Care Unit}
  \item{óbitos}{deaths}
  \item{hist_desloc}{sex}
  \item{dt_is}{sex}
  \item{tp_gestao}{Management Kind}
  \item{tipo_municipal}{Municipal Management}
}}
\source{
\href{http://painel.saude.rj.gov.br/monitoramento/covid19.html}{Secretaria de Saúde - RJ}
}
\usage{
coronavirus_rj_case_metadata
}
\description{
Metadata of the Rio de Janeiro COVID-19 cases
}
\details{
Metadata of the Rio de Janeiro COVID-19 cases
}
\examples{

data(coronavirus_rj_case_metadata)

}
\keyword{COVID19}
\keyword{coronavirus_br}
\keyword{datasets}
