% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coronavirusbrazil-package.R
\docType{package}
\name{coronavirusbrazil-package}
\alias{coronavirusbrazil}
\alias{coronavirusbrazil-package}
\title{coronavirusbrazil: The 2019 Novel Coronavirus COVID-19 (2019-nCoV) Dataset for Brazil}
\description{
Daily updated dataset for coronavirus (covid-19) cases for Brazil.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mralbu/coronavirusbrazil}
  \item Report bugs at \url{https://github.com/mralbu/coronavirusbrazil/issues}
}

}
\author{
\strong{Maintainer}: Marcelo Albuquerque \email{marceloralbuquerque@gmail.com}

}
\keyword{internal}
