% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_covid.R
\name{plot_coronavirus}
\alias{plot_coronavirus}
\title{plot_coronavirus}
\usage{
plot_coronavirus(
  coronavirus_br,
  xaxis = "date",
  yaxis = "cases",
  color = NULL,
  log_scale = FALSE,
  linear_smooth = FALSE,
  filter_variable = NULL,
  filter_values = NULL,
  facet = NULL
)
}
\arguments{
\item{coronavirus_br}{Dataset}

\item{xaxis}{xaxis variable ("cases" | "date" | "days_gt_10" | "days_gt_100")}

\item{yaxis}{yaxis variabel ("cases" | "deaths" | "new_cases" | "new_deaths" | "death_rate")}

\item{color}{coloring data.frame variable}

\item{log_scale}{plot on log scale?}

\item{linear_smooth}{plot line linear smooth?}

\item{filter_variable}{variable to be filtered}

\item{filter_values}{filtering values}

\item{facet}{variable to be used for facetting}
}
\value{

}
\description{
plot_coronavirus
}
