% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_covid.R
\name{plot_coronavirus_br}
\alias{plot_coronavirus_br}
\title{plot_coronavirus_br}
\usage{
plot_coronavirus_br(
  coronavirus_br,
  xaxis_br = "date",
  what_br = "cases",
  delta_br = FALSE,
  log_scale_br = FALSE,
  tendency_br = FALSE
)
}
\arguments{
\item{coronavirus_br}{Dataset}

\item{xaxis_br}{xaxis variable}

\item{what_br}{yaxis variabel}

\item{delta_br}{plot variation?}

\item{log_scale_br}{plot on log scale?}

\item{tendency_br}{plot line tendency?}
}
\value{

}
\description{
plot_coronavirus_br
}
