% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_covid.R
\name{plot_coronavirus_states}
\alias{plot_coronavirus_states}
\title{plot_coronavirus_states}
\usage{
plot_coronavirus_states(
  coronavirus_br_states,
  xaxis = "date",
  what = "cases",
  filter_state = c("RJ", "SP"),
  delta = TRUE,
  log_scale = TRUE,
  facet_state = TRUE,
  tendency = FALSE
)
}
\arguments{
\item{xaxis}{xaxis variable}

\item{what}{yaxis variabel}

\item{filter_state}{select states to plot}

\item{delta}{plot variation?}

\item{log_scale}{plot on log scale?}

\item{facet_state}{facet by state?}

\item{tendency}{plot line tendency?}

\item{coronavirus_br}{Dataset}
}
\description{
plot_coronavirus_states
}
