% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{spatial_br_cities}
\alias{spatial_br_cities}
\title{Spatial Dataset of Brazilian Cases by state}
\format{}
\source{
\href{https://brasil.io/dataset/covid19/}{brasil.io}
}
\usage{
spatial_br_cities
}
\description{
Spatial summary of the Coronavirus (COVID-19) cases in Brazilian states
}
\details{
The spatial dataset contains cases of covid-19 in Brazilian cities
}
\examples{

data(spatial_br_cities)

}
\keyword{COVID19}
\keyword{coronavirus_br}
\keyword{datasets}
