% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{spatial_br_states}
\alias{spatial_br_states}
\title{Spatial Dataset of Brazilian Cases by state}
\format{}
\source{
\href{https://covid.saude.gov.br}{Ministerio da Saude}
}
\usage{
spatial_br_states
}
\description{
Spatial summary of the Coronavirus (COVID-19) cases in Brazilian states
}
\details{
The spatial dataset contains cases of covid-19 in Brazilian states
}
\examples{

data(spatial_br_states)

}
\keyword{COVID19}
\keyword{coronavirus_br}
\keyword{datasets}
