% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{spatial_rj_neighborhoods}
\alias{spatial_rj_neighborhoods}
\title{Spatial Dataset of Rio de Janeiro cases by neighborhood}
\format{A spatial data.frame object (sf)
\describe{
  \item{neighborhood}{Rio de Janeiro Neighborhood}
  \item{cases}{Confirmed Cases}
  \item{lat}{Latitude}
  \item{lon}{Longitude}
  \item{AP}{AP}
  \item{objectId}{objectId}
  \item{geoms}{Simple Features geometries}
}}
\source{
\href{http://painel.saude.rj.gov.br/monitoramento/covid19.html}{Secretaria de Saúde - RJ}
}
\usage{
spatial_rj_neighborhoods
}
\description{
Spatial summary of the Coronavirus (COVID-19) cases in Rio de Janeiro neighborhoods
}
\details{
The spatial dataset contains cases of covid-19 in Rio de Janeiro neighborhoods
}
\examples{

data(spatial_rj_neighborhoods)

}
\keyword{COVID19}
\keyword{coronavirus_br}
\keyword{datasets}
