% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{spatial_world}
\alias{spatial_world}
\title{Spatial Dataset of World Cases}
\format{}
\source{
\href{https://github.com/RamiKrispin/coronavirus}{RamiKrispin/coronavirus}
}
\usage{
spatial_world
}
\description{
Spatial summary of the Coronavirus (COVID-19) world cases
}
\details{
The spatial dataset contains world cases of covid-19
}
\examples{

data(spatial_world)

}
\keyword{COVID19}
\keyword{coronavirus_br}
\keyword{datasets}
