% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_datasets.R
\name{update_datasets}
\alias{update_datasets}
\title{Update Datasets}
\usage{
update_datasets(filename = NULL)
}
\description{
Update coronavirus_br, coronavirus_br_states, coronavirus_br_cities, spatial_br_states, spatial_br_cities datasets
 Data Sources: https://coronavirus.saude.gov.br/ and https://brasil.io/dateset/coronavirus19/
}
\examples{
\dontrun{
update_datasets()
}
}
